+++ be/src/column/variant_merger.h
@@ -14,37 +14,13 @@
 
 #pragma once
 
-#include <optional>
-#include <vector>
-
 #include "base/status.h"
 #include "base/statusor.h"
 #include "column/column.h"
-#include "column/variant_path_parser.h"
-#include "types/variant_value.h"
 
 namespace starrocks {
 
 class VariantColumn;
-
-struct VariantOverlayInput {
-    VariantPath path;
-    VariantRowValue value;
-};
-
-// VariantMergerPolicy keeps merge-time type arbitration and row overlay rebuild
-// in one place so reader-side and merge-side prefix handling can share the same
-// rules without coupling their orchestration code.
-class VariantMergerPolicy {
-public:
-    static StatusOr<TypeDescriptor> choose_common_type(const TypeDescriptor& lhs, const TypeDescriptor& rhs);
-
-    static StatusOr<MutableColumnPtr> cast_typed_column(const Column& src_col, const TypeDescriptor& src_type_desc,
-                                                        const TypeDescriptor& dst_type_desc);
-
-    static StatusOr<VariantRowValue> build_row_from_overlays(std::optional<VariantRowRef> base,
-                                                             std::vector<VariantOverlayInput> overlays);
-};
 
 // VariantColumnMerger centralizes VariantColumn merge orchestration.
 //
@@ -73,6 +49,11 @@ public:
 // to expand in later stages.
 class VariantColumnMerger {
 public:
+    static StatusOr<TypeDescriptor> choose_common_type(const TypeDescriptor& lhs, const TypeDescriptor& rhs);
+
+    static StatusOr<MutableColumnPtr> cast_typed_column(const Column& src_col, const TypeDescriptor& src_type_desc,
+                                                        const TypeDescriptor& dst_type_desc);
+
     // Reconcile overlapping shredded path type conflicts between dst and src.
     // Numeric/decimalv3 paths are widened when possible; unsupported conflicts are hoisted to TYPE_VARIANT.
     static Status arbitrate_type_conflicts(VariantColumn* dst, VariantColumn* src);
